+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "vscode-ui5" extension will be documented in this fil
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+- Added snippet for Title (ui5-d-ttl_a, and ui5-d-ttl)
 
 ## [1.0.0]
 - All snippets use current filename for naming purposes.